requirements.txt
gradio
mistralai
pathspec
python_dotenv